refactor(pfb): move print settings into a dialog behind a gear icon
